Refactor NewTask and TaskDetails page components to organize better and fix mistakes
