+++ src/pages/TaskDetails/TaskDetailsDocker.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const TaskDetailsDocker = () => {
+  return <div></div>;
+};
+
+export default TaskDetailsDocker;
